change search logic to make arguments optional
WallHavenAPI.search() :
Some of if statements raise exceptions for options even if they are None
so they're not optional at all and user is enforced to use them.
I fixed this, so the script checks if some option is valid only when it is
passed to the method.

Moreover, some of if statements can be easily converted from:
if Obj is not None:
    ...
To:
if Obj:
    ...
Which is more clear logically.

Also there are some trailing space deletion and code formatting along the way.