fix "purity" and "category" filters
Wallhaven.cc API is using 3 digits (000) filters where each digit represents
some of the filters.

For example:
    (general,anime,people = 111)
    (general = 100)
    (anime   = 010)
    (people  = 001)

But the previous code doesn't work exactly as described.
For example it interprets people as "111" value (all filters are enabled).

So I wrote 2 methods, which accepts list of filters (e.g. ["sfw", "sketcy"])
and return filter value (e.g. "110"), so now we can pass more than one
filter and it will work perfectly. But if some of the filters are unknown
the script still throw an exception.